requirements-test.txt
pytest>=2.6.4
pytest-cov>=1.8.1
pytest-django>=2.8.0
pytest-remove-stale-bytecode>=1.0
# pytest-random>=0.2
pytest-sugar>=0.4.0
# pytest-spec>=0.2.24
celery>=3.1.18
pinax-eventlog>=1.0.0